Fix float64-config dtype crashes; always use configured dtype
Audit cluster 1 (TORCHREF_AUDIT.md): several paths hardcoded float32 /
complex128 / .float() in tensors that interact with data tensors carrying
the configured dtype. Under a float64 config (dtypes.float=float64) these
raised dtype mismatches in scatter_add / matmul, or silently downcast.

Crash fixes (high severity):
- scaler_base.get_binwise_mean_intensity: scatter_add accumulators now
  match fobs.dtype instead of the default float32.
- collection_scaler._calc_initial_scale_joint: scales/counts use
  get_float_dtype() so scatter_add_ against fobs-derived log_ratios works.
- occupancy_floor_diagnostic: hkl.T.float() -> hkl.T.to(positions dtype)
  in compute_density_at_positions and NegativeDensityPenalty.forward.

Consistency fixes (low severity):
- sf_ds: complex128 accumulators -> get_complex_dtype(); equiv_hkls .float()
  -> .to(self.dtype_float).
- base/reciprocal/symmetry.compute_translation_phases: drop the float32 cast
  (phase already carries the configured float dtype).
- mixed_model / model_collection: fraction tensors use the base models'
  float dtype instead of hardcoded float32.

Adds tests/integration/test_dtype_config_float64.py exercising these paths
under a float64 config (would crash pre-fix).

+++ torchref/refinement/targets/occupancy_floor_diagnostic.py
@@ -97,7 +97,9 @@ class OccupancyFloorDiagnostic:
             # Compute h·r for all position-reflection pairs
             # positions: (N, 3), hkl: (M, 3)
             # h_dot_r: (N, M)
-            h_dot_r = torch.matmul(positions, hkl.T.float())
+            # Match hkl to the positions' (configured) dtype so the matmul
+            # does not raise under a float64 config.
+            h_dot_r = torch.matmul(positions, hkl.T.to(dtype=positions.dtype))
 
             # Fourier sum: ρ(r) = Σ_h F(h) * exp(2πi * h·r)
             # For real density, this is: Σ_h |F(h)| * cos(2π*h·r + φ(h))
@@ -343,8 +345,11 @@ class NegativeDensityPenalty(DeviceMixin, nn.Module):
         # Get mixed model structure factors (includes the α weighting)
         fcalc_mixed = self.mixed_model(self.hkl, recalc=True)
 
-        # Compute density at dark atom positions
-        h_dot_r = torch.matmul(self.frac_positions, self.hkl.T.float())
+        # Compute density at dark atom positions (match hkl to the
+        # configured dtype of frac_positions so float64 does not raise)
+        h_dot_r = torch.matmul(
+            self.frac_positions, self.hkl.T.to(dtype=self.frac_positions.dtype)
+        )
         phase = torch.angle(fcalc_mixed)
         amplitude = torch.abs(fcalc_mixed)
